commit-graph.h: store an odb in 'struct write_commit_graph_context'
In a previous patch, the 'char *object_dir' in 'struct commit_graph' was
replaced with a 'struct object_directory'. This patch applies the same
treatement to 'struct write_commit_graph_context', which is an
intermediate step towards getting rid of all path normalization in
'commit-graph.c'.

Instead of taking a 'char *object_dir', functions that construct a
'struct write_commit_graph_context' now take a 'struct object_directory
*'. Any code that needs an object directory path use '->path' instead.

This ensures that all calls to functions that perform path normalization
are given arguments which do not themselves require normalization. This
prepares those functions to drop their normalization entirely, which
will occur in the subsequent patch.

Signed-off-by: Taylor Blau <me@ttaylorr.com>