fixup! pack-bitmap.c: boundary-based walk
If we have haves list, it doesn't matter whether or not the haves
themselves are in the bitmapped pack (or MIDX), but whether anything
between us and the boundary is in the bitmap.

We could defer this check until after we have discovered what that
boundary is, but we could also just fulfill the request. This may be
worth revisiting later (i.e., should we fall back to the non-bitmap
traversal if none of the boundary is in the bitmap?)