fixup! pack-bitmap.c: boundary-based walk
Clearing the UNINTERESTING bit from all parsed objects is heavy-handed,
and causes us to spend much more time generating a bitmap for the haves
side then if we only clear the UNINTERESTING bits from commits that we
know we just marked UNINTERESTING in the traversal above (when we
discovered what the boundary was).

This seems to make us go quite a bit faster, but I'm not sure I fully
understand why. It also means that we'll *always* include an extra SLOP
commits in the output. Not sure what to do about that...