pack-bitmap: skip looking for bitmaps of redundant objects
They can't possibly contribute if they're already in the bitmap. The
existing find_objects() code _might_ want to use the same optimization,
but it's a bit of a tradeoff. We spend extra work to find the bit
position of each object via the revindex, but save ourselves looking up
the commit in the bitmap hash. That lookup is cheap-ish, but if we
expect to find a lot of redundant objects, then actually OR-ing their
bitmaps together is surprisingly expensive.